+++ src/firmware/application/io/buttons/stub/Filter.h
@@ -27,7 +27,7 @@ namespace io
         public:
         ButtonsFilter() = default;
 
-        bool isFiltered(size_t index, uint8_t& numberOfReadings, uint32_t& states) override
+        bool isFiltered(size_t index, uint8_t& numberOfReadings, uint16_t& states) override
         {
             return false;
         }
