+++ tests/include/tests/stubs/HWAEncoders.h
@@ -7,5 +7,5 @@ class HWAEncoders : public io::Encoders::HWA
     public:
     HWAEncoders() = default;
 
-    MOCK_METHOD3(state, bool(size_t index, uint8_t& numberOfReadings, uint32_t& states));
+    MOCK_METHOD3(state, bool(size_t index, uint8_t& numberOfReadings, uint16_t& states));
 };
